BGA-116 Improve celo transaction builder test suite
